Add files via upload
This code creates an object called person that has properties for the person's name, age, and job. It then logs a message to the console that uses the object's properties to construct a sentence describing the person. The code then modifies the job property of the person object and adding a new property hobby and then logs a new message that again uses the properties of the object to construct a sentence, this time describing the person's updated job and hobby.