Add files via upload
This code prompts the user for their name and age using the "prompt()" function, and stores the input in the "userInput" and "age" variables respectively. It then uses "console.log()" to display a greeting and the user's age. Next, it creates an array of fruits and prompts the user to enter a fruit. It checks whether the user's input is in the list of fruits using the "indexOf()" function, and displays a message accordingly.