Add files via upload
The program defines a function called smallestCommonMultiple() which does not take any parameter. It prompts the user for input using the prompt() function and assigns the input to two variables called num1 and num2.

The function then declares three variables multiple, i, and j and sets multiple to 0 and i to 1.
It uses a while loop that iterates until multiple is not equal to 0. Inside the while loop, it calculates the multiples of num1 and num2 by using the i variable.
If the multiples are the same, it assigns it to the multiple variable and break the loop.
It increments the i by 1 at the end of each iteration.

Finally the function returns the multiple variable. The program also includes a console.log statement that prints the smallest common multiple of the two numbers entered by the user.