Add files via upload
The program defines a function called fibonacci() which prompts the user for input using the prompt() function and assigns the input to a variable called num.
It declares an array fib and initialize it with the first two numbers of the Fibonacci sequence.
The function then uses a while loop that iterates until the length of the fib array is less than num.
Inside the while loop, it calculates the next number in the Fibonacci sequence by adding the last two numbers in the fib array.
It then pushes this new number to the fib array.

Once the while loop terminates, the function returns the fib array joined by commas. The program also includes a console.log statement that prints the Fibonacci sequence up to the number entered by the user.