Add files via upload
This program takes two user inputs, which are two numbers. When the user clicks the "Calculate" button, the program converts the input from string to float, and then performs an addition of those two numbers, assigns the result to variable called result, and then logs the result in the console. This program uses JavaScript's parseFloat() function to convert a string to a floating-point number, and the console.log() method to display output in the browser's developer console.