Add files via upload
This program uses the prompt() function to get user input, and parseInt() function to convert the input from string to integer. Then it uses a for loop to check if the number is divisible by any number between 2 and the square root of the number. If the number is divisible by any of these numbers, it means it's not prime. If the number is not divisible by any of these numbers, it means it's prime.

Please note that this program uses a simple algorithm to check if the number is prime or not, and it may not be efficient for very large numbers. Also this code should be run in a JavaScript environment like a web browser or a Node.js environment.