Add files via upload
This program is a shorter version of a class that represents a BankAccount. The class has a constructor that takes the name and the balance of the account, and a deposit method that increases the balance by the amount passed in. The class doesn't have any withdraw or check balance method in this version, but you can add them easily as per your requirement. The program creates an instance of the class, and performs a deposit transaction.