Add files via upload
javascript
Copy code
class Employee {
  constructor(name, salary) {
    this.name = name;
    this.salary = salary;
  }
  
  work() {
    console.log(`${this.name} is working.`);
  }
  
  getPaid() {
    console.log(`${this.name} has been paid ${this.salary}.`);
  }
}

class Manager extends Employee {
  constructor(name, salary, department) {
    super(name, salary);
    this.department = department;
  }
  
  work() {
    console.log(`${this.name} is managing the ${this.department} department.`);
  }
  
  assignTask(employee, task) {
    console.log(`${this.name} assigned ${task} to ${employee.name}.`);
    employee.work();
  }
}

const jim = new Employee("Jim", 50000);
const pam = new Employee("Pam", 40000);
const dwight = new Manager("Dwight", 100000, "Sales");

dwight.assignTask(jim, "make sales call");
dwight.assignTask(pam, "create a sales presentation");
This program demonstrates the use of JavaScript classes and inheritance. It defines an Employee class with a constructor that takes the name and salary of an employee. It also has two methods work() and getPaid() which logs a message when invoked.

It also defines a Manager class that inherits from the Employee class, and it has its own constructor that takes the name, salary, and department of the manager. The Manager class overrides the work() method to show that it's managing the department and has a new method assignTask(employee, task) which takes an employee and a task and logs a message that the task has been assigned to that employee and then the employee will work on that task.

In the main part of the code, it creates an instance of Employee and Manager class, assigns tasks to employees and manager work method will be called.