Add files via upload
This program is an example of implementing a DoubleEndedQueue class in JavaScript, which is a data structure that allows items to be added or removed from both ends. It demonstrates the use of classes, inheritance, and methods for handling the various DoubleEndedQueue operations such as adding and removing items from the beginning and end of the queue, as well as checking the front item of the queue and the current items present in it.